# backend/docs/swagger.yaml
basePath: /
definitions:
...
  request.Request:
    properties:
      createdAt:
        type: string
      id:
        type: string
      roomIcon:
        type: string
      roomId:
        type: string
      roomName:
        type: string
      senderAvatar:
        type: string
      senderCode:
        type: string
      senderId:
        type: string
      senderName:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: alisfrk2004@gmail.com
    name: Дарья
  description: Серверная часть веб-приложения для коммуникации.
  termsOfService: http://swagger.io/terms/
  title: Buzz API
  version: "1.0"
paths:
...
  /requests:
    get:
      description: Возвращает список ожидающих запросов (в друзья и приглашения в
        комнаты) для текущего пользователя. Ответ содержит поля, зависящие от типа
        запроса.
      produces:
      - application/json
      responses:
        "200":
          description: Список входящих запросов (может быть пустым)
          schema:
            items:
              $ref: '#/definitions/request.Request'
            type: array
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить все входящие запросы
      tags:
      - requests
  /requests/{id}/accept:
    post:
      description: Принимает запрос (в друзья или приглашение в комнату) по его ID.
        После принятия пользователь становится другом или участником комнаты.
      parameters:
      - description: ID запроса
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Запрос принят
        "400":
          description: Неверный ID запроса или запрос не в статусе pending
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Запрос не принадлежит текущему пользователю
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Запрос не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Запрос уже был обработан (не pending)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Принять запрос
      tags:
      - requests
  /requests/{id}/reject:
    post:
      description: Отклоняет запрос (в друзья или приглашение в комнату) по его ID.
        Запрос удаляется.
      parameters:
      - description: ID запроса
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Запрос отклонён
        "400":
          description: Неверный ID запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Запрос не принадлежит текущему пользователю
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Запрос не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отклонить запрос
      tags:
      - requests
  /requests/outgoing:
    get:
      description: Возвращает список отправленных запросов (в друзья и приглашения
        в комнаты) для текущего пользователя. Ответ содержит поля, зависящие от типа
        запроса.
      produces:
      - application/json
      responses:
        "200":
          description: Список исходящих запросов (может быть пустым)
          schema:
            items:
              $ref: '#/definitions/request.Request'
            type: array
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить все исходящие запросы
      tags:
      - requests
...
securityDefinitions:
  BearerAuth:
    description: Введите токен в формате "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"